+++ src/brain/index.ts
@@ -2,7 +2,7 @@ import { containerBootstrap } from "@nlpjs/core";
 import { Nlp } from "@nlpjs/nlp";
 import { LangEn } from "@nlpjs/lang-en-min";
 import { removeEmojis } from "@nlpjs/emoji";
-import { Reply } from "../types/Reply";
+import { BrainReply } from "../types/BrainReply";
 import prayerCorpus from "./en/prayer";
 import mainCorpus from "./en/main";
 import greetingsCorpus from "./en/greetings";
@@ -31,19 +31,21 @@ async function trainBrain(): Promise<any> {
   nlp.onIntent = onIntent;
 
   await nlp.train();
+
+  atomStore.set(brainAtom, nlp);
   return nlp;
 }
 
-async function getBrainReply(q: string): Promise<Reply> {
+async function getBrainReply(q: string): Promise<BrainReply> {
   const brain = atomStore.get(brainAtom);
 
   if (!brain)
     throw new Error("You should train your brain first to get a reply.");
 
-  const brainResponse = await brain.process("en", removeEmojis(q));
-  console.log(brainResponse);
+  const brainReply = await brain.process("en", removeEmojis(q));
+  console.log(brainReply);
 
-  return brainResponse;
+  return brainReply;
 }
 
 export { trainBrain, getBrainReply };
