+++ src/components/BottomNav.tsx
@@ -27,7 +27,7 @@ const BottomNav: React.FC<BottomNavProps> = () => {
   const router = useIonRouter();
 
   return (
-    <div className="btm-nav">
+    <div className="btm-nav duration-300">
       {navs.map((nav, index) => (
         <IonRouterLink
           className={`text-current ${
