+++ src/constants/store.ts
@@ -1,1 +1,4 @@
+/**
+ * localStorage key for the theme
+ */
 export const THEME_STORE_KEY = "theme";
