+++ src/helpers/string.ts
@@ -1,3 +1,7 @@
+/**
+ * html decode
+ * @example htmlEncode('<div>content</div>') // content
+ */
 export function htmlDecode(input: string) {
   var doc = new DOMParser().parseFromString(input, "text/html");
   return doc.documentElement.textContent;
